Refactor: simplified Explore and ExploreView code
Explore now only acts as an object to fetch the json, ExploreView
does the check for the required data live. (it would have done it
anyway to check if the std::optional had a value or not).
Data initialization has been delegated to the data fetching thread
the main thread only draws the widgets now.