Fixed segfault/memory leak
There was a bug that was causing a segfault during application closure
which I temporarily worked aroud by using a raw C pointer, allocating
the resource and then forget about it (voluntary memory leak).

The came from the fact that some widgets need a pointer to their parent
toplevel (parent Window) which is the MainWindow object, and which is
(for now) instantiated automatically by GTK. This widget then contains
a unique_ptr to ExploreView, which internally contains a shared_ptr to
the parent Window.
This means that MainWindow is created outside of ExploreView, contains
ExploreView and then passes it's pointer to ExploreView which puts it
in a smart pointer. When ExploreView goes auto of scope it dutifully
deallocates the shared_ptr and kills MainWindow, which was just going
out of scope. This results in a double free and therefore a segfault
during application closure.